refactor: (event-detail) EventDetail 컴포넌트 분리 및 데이터 페칭 역할 분리
